refactor: move balance update logic from TransactionObserver to new BalanceUpdateService
